Initial Climaduino web controller code
This code is mainly feature complete. It interfaces with the Climaduino
Controller that controls the AC system and detects temperature and
humidity. Parameters can be changed and historical data viewed from the
web interface that is optimized for mobile. Programmable setpoints are
also available, but must be programmed in using the /admin pages. It
would be ideal to have a front-end page for this at some point.

There is no security at this point.

This is my first attempt using the Django framework, so I may have done
a lot of things wrong.

I welcome your input and fixes for this code.