Controller robustness, noted an unfixed bug
Made the controller component more robust in the face of rrd database
update failures. These failures should not happen in the field, but
happened in my testing when trying to isolate the cause of another bug
and changing system time.

Bug exists where when setting a program for midnight (00:00) it does
not work. Added this into the source code, but bug still appears to
exist.

Updated programming sentry to use datetime.time as earliest time.
Probably no functional difference since Django database query was
probably ignoring the date component, but it seems better to strip out
the date information regardless.