Climaduino controller serial robustness
Made the climaduino controller code try to reconnect the serial
connection on error. That way if the Climaduino thermostar is unplugged
from the web controller, it will start to work again if it is plugged
back in rather than erroring out and requiring a restart of the
Climaduino Web controller software.